+++ tools/web_repl.mjs
@@ -60,11 +60,11 @@ export default function(compiler, baselib) {
             return true;
         },
 
-        'compile': function web_repl_compile(code, opts) {
+        'compile': async function web_repl_compile(code, opts) {
             opts = opts || {};
             opts.keep_docstrings = true;
             opts.filename = '<input>';
-            return streaming_compiler.compile(code, opts);
+            return await streaming_compiler.compile(code, opts);
         },
 
         'runjs': function runjs(code) {